Fixed set face not updating hud, copying map tiles crash, and pathfinding error with event move toward player routes
Former-commit-id: 99ba8eeaed0d9ba482dd458be40c291dafcd1d08

+++ Intersect.Server/Classes/Entities/EventPageInstance.cs
@@ -65,8 +65,8 @@ namespace Intersect.Server.Entities
             Id = BaseEvent.Id;
             MyPage = myPage;
             MapId = mapId;
-            X = (byte)eventIndex.X;
-            Y = (byte)eventIndex.Y;
+            X = eventIndex.X;
+            Y = eventIndex.Y;
             Name = myEvent.Name;
             MovementType = MyPage.Movement.Type;
             MovementFreq = MyPage.Movement.Frequency;
